# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TripPlanner API",
        "description": "Complete REST API for TripPlanner backend with NAVER Cloud Platform integration",
        "contact": {
            "email": "support@tripplanner.example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/places/search": {
            "post": {
                "tags": [
                    "Places"
                ],
                "summary": "Search for places by text query",
                "description": "Search for places using text query.",
                "operationId": "f023a92109ab300a59ac9e7e49de48ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "query"
                                ],
                                "properties": {
                                    "query": {
                                        "type": "string",
                                        "example": "Tokyo Tower"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Search results from NAVER Places",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Place"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "query": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "503": {
                        "description": "NAVER API service unavailable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/places/search-nearby": {
            "post": {
                "tags": [
                    "Places"
                ],
                "summary": "Search for places near specific coordinates",
                "description": "Search for nearby places by coordinates.",
                "operationId": "ba199461499f0aa9e490914ae5c82bc9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "latitude": {
                                        "type": "number",
                                        "format": "double",
                                        "example": 35.6762
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "double",
                                        "example": 139.6503
                                    },
                                    "query": {
                                        "type": "string",
                                        "example": "restaurant"
                                    },
                                    "radius": {
                                        "description": "Search radius in meters (default: 500)",
                                        "type": "integer",
                                        "example": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nearby places from NAVER Local Search",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Place"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "search_location": {
                                                    "properties": {
                                                        "latitude": {
                                                            "type": "number"
                                                        },
                                                        "longitude": {
                                                            "type": "number"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "radius": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/places/naver/{naverPlaceId}": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "Get place details by NAVER place ID",
                "description": "Get place details by NAVER place ID.",
                "operationId": "4e4cdbcd0d9bf596283a5eec617cf28b",
                "parameters": [
                    {
                        "name": "naverPlaceId",
                        "in": "path",
                        "description": "NAVER place ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Place details from NAVER Maps",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Place"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/places": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "List all saved places from database",
                "description": "Display a listing of saved places.",
                "operationId": "98b7b5e708ea1544e70cf25ec8621387",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Results per page (default: 15)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of saved places with reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Place"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Places"
                ],
                "summary": "Save a place from NAVER to database",
                "description": "Store a place from NAVER to database.",
                "operationId": "9b1406d5892a139db93bd35d7a248450",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naver_place_id"
                                ],
                                "properties": {
                                    "naver_place_id": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "fetch_details": {
                                        "description": "Fetch full details from NAVER (default: false)",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "name": {
                                        "description": "Used only if fetch_details is false",
                                        "type": "string",
                                        "example": "Tokyo Tower"
                                    },
                                    "latitude": {
                                        "description": "Used only if fetch_details is false",
                                        "type": "number",
                                        "format": "double",
                                        "example": 35.6586
                                    },
                                    "longitude": {
                                        "description": "Used only if fetch_details is false",
                                        "type": "number",
                                        "format": "double",
                                        "example": 139.7454
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Place saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Place saved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Place"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Place already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Place already exists"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Place"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Place": {
                "title": "Place",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "naver_place_id": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "name": {
                        "type": "string",
                        "example": "Tokyo Tower"
                    },
                    "address": {
                        "type": "string",
                        "example": "4 Chome-2-8 Shibakoen, Minato City, Tokyo"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double",
                        "example": 35.6586
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double",
                        "example": 139.7454
                    },
                    "category": {
                        "type": "string",
                        "example": "Landmark"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+81-3-3433-5111",
                        "nullable": true
                    },
                    "rating": {
                        "type": "number",
                        "format": "double",
                        "example": 4.5,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "Unauthorized": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UnauthorizedError"
                        }
                    }
                }
            },
            "ValidationError": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
...
            "NotFound": {
                "description": "Resource not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/NotFoundError"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication. Use: Bearer {your-token}",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}